Converterの取得関連をutil化
